Funcionamiento del juego

+++ AdivinaDondeEstoyProyecto/app/src/main/java/Modelo/Leyend.kt
@@ -1,3 +1,3 @@
 package Modelo
 
-class Leyend (var nombre: String, var cordenadaX: Float, var cordenadaY: Float, var historia: String, var acertado: Boolean)
+class Leyend (var nombre: String, var cordenadaX: Double, var cordenadaY: Double, var historia: String, var acertado: Boolean)
